Method for saving vocabulary into a file added and tested. Should be working. Bug at TranslationUnit::operator>> fixed, weight was not stored.
The application is not yet saving the vocabulary, since it is not modifying it and the functionality is not yet implemented.

+++ src/TranslationUnit.h
@@ -23,13 +23,25 @@ class TranslationUnit {
     TranslationUnit(const TranslationUnit & other) = default;
     TranslationUnit(TranslationUnit && other) = default;
 
+    /**
+     * This function saves data into a given stream. The data stored have this syntax:
+     *  weight;to_translate;translation[;translation[;translation...]]
+     * where each word (text in between semicolons) has to contain only letters, dashes, apostrophes, or spaces. The word
+     * has not to be empty or only consist of spaces.
+     */
     friend std::ostream & operator<<(std::ostream & os, const TranslationUnit & unit);
 
+    /**
+     * This function loads data from a given stream. Data in the stream needs to have this syntax:
+     *  weight;to_translate;translation[;translation[;translation...]]
+     * where word to translate and translation has to contain only letters, dashes, apostrophes, or spaces. The word
+     * has not to be empty or only consist of spaces.
+     */
     friend std::istream & operator>>(std::istream & is, TranslationUnit & unit);
 
     std::string m_word_to_translate;
     std::vector<std::string> m_translation;
-    unsigned m_weight;
+    size_t m_weight;
 
   private:
     /**
